Added databases for user content: avatars, maps, props which include some usage of Elixir macros to provide common data. Also added database shards (V-Sekai server instances) and events with start and end dates, and a many to many database for friendships between users. Users now has an is_admin boolean which gives access to several exclusive pages which allow global unrestricted viewing and management of the avatar, maps, props, and shard databases. Fixed warnings on function for user credentials validation, as well as a total refactor of both the HTML and API/v1 SessionControllers which moves a lot of reusable code into auth.ex helper class. A Discord PowAssent provider was also added.
The new databases are mostly fore-planning for functionality we want to add to groups including user content uploading, user hosting of shards, and friending other users. For now though, the intention is to allow a user or users designated as admins to manage these databases through the website. The improvements to the credentials validation code was to fix some warnings and to make the code more readable and reusable. The Discord provider was added with the intention of providing OAuth support provided by Discord along with the planned Google login, though currently it is not set up as usable.

To test these new features, it is currently necessary to register for an account, but since there is currently know way to designate a user as an admin, this boolean must be set via an external database tool. Once active though and when logged in as an admin, the paths '/admin/avatars', '/admin/maps', '/admin/props', and '/admin/shards' can be accessed and edited. In order to test login from V-Sekai, launch the application as an editor an click on the Uro button in the top right-hand corner. Log-in with your credentials and if successful, an ini file containing a login token will be stored in your users:// folder.